requirements.txt
alembic==1.14.1
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
croniter==6.0.0
cryptography==44.0.2
debtcollector==3.0.0
decorator==5.2.1
distro==1.9.0
Django==5.1.4
django-debug-toolbar==4.4.6
dnspython==2.7.0
docx2txt==0.9
dogpile.cache==1.3.4
easyocr==1.7.2
eval_type_backport==0.2.2
eventlet==0.39.0
fastapi==0.115.12
fasteners==0.19
filelock==3.18.0
filetype==1.2.0
fsspec==2025.3.2
futurist==3.1.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.68.0
greenlet==3.1.1
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
iso8601==2.1.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keystoneauth1==5.10.0
keystonemiddleware==10.9.0
kombu==5.4.2
langchain==0.3.23
langchain-core==0.3.54
langchain-google-genai==2.0.10
langchain-groq==0.2.5
langchain-text-splitters==0.3.8
langsmith==0.3.32
lazy_loader==0.4
legacy-cgi==2.6.2
logutils==0.3.5
Mako==1.3.9
MarkupSafe==3.0.2
mistral==19.0.0
mistral-lib==3.3.1
mistralai==1.5.1
mpmath==1.3.0
msgpack==1.1.0
mypy-extensions==1.0.0
narwhals==1.30.0
netaddr==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.3
openai==1.75.0
opencv-python-headless==4.11.0.86
orjson==3.10.16
os-service-types==1.7.0
oslo.cache==3.10.1
oslo.concurrency==7.1.0
oslo.config==9.7.1
oslo.context==5.7.1
oslo.db==17.2.1
oslo.i18n==6.5.1
oslo.log==7.1.0
oslo.messaging==16.1.0
oslo.metrics==0.11.0
oslo.middleware==6.3.1
oslo.policy==4.5.1
oslo.serialization==5.7.0
oslo.service==4.1.1
oslo.utils==8.2.0
osprofiler==4.2.0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.1
pdfminer.six==20250327
pdfplumber==0.11.6
pecan==1.5.1
pillow==11.2.1
ply==3.11
prettytable==3.15.1
prometheus_client==0.21.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycadf==4.0.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
PyJWT==2.10.1
PyMuPDF==1.25.5
PyNaCl==1.5.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-keystoneclient==5.6.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
razorpay==1.4.2
referencing==0.36.2
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
setuptools==75.8.2
shapely==2.1.0
simplegeneric==0.8.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
statsd==4.0.1
stevedore==5.4.1
streamlit==1.43.1
sympy==1.14.0
tenacity==9.1.2
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
tifffile==2025.3.30
toml==0.10.2
tooz==6.3.0
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
WebOb==1.8.9
wrapt==1.17.2
WSME==0.12.1
yappi==1.6.10
yaql==3.0.0
zipp==3.21.0
zstandard==0.23.0
